requirements.txt
dj-static==0.0.6
Django==1.10.3
DjangoRestless==0.0.10
flickrapi==2.1.2
gunicorn==19.6.0
httplib2==0.9.2
oauthlib==2.0.0
python-social-auth==0.2.21
python3-openid==3.0.10
requests==2.3.0
requests-oauthlib==0.7.0
requests-toolbelt==0.7.0
simplejson==3.10.0
six==1.7.3
spotipy==2.3.8
static==1.1.1
static3==0.7.0
virtualenv==15.0.3
whitenoise==3.2.2
